* master:
  Remove configure option --max-api-stats which does not do anything. Should have been removed as part of commit ea1fb0c87261b1fbf375fa6ade26deda1d01995b
  Add extension ip.test.ext to Au Test, with Test method to allocate extra TCP ports.
  Optimize HTTPHdr conversion of HTTP/2 to HTTP/1.1
  Return TSFetchSM from TSFetchUrl so TSFetchFlagSet can set fetch flags
  READ_RESPONSE_HDR_HOOK is invoked only when the response is from the Origin Server
  Fix typo in system stats, change loadavg 10min to be 15min (#6608)
  Making client session id unique across HTTP/1 and 2 sessions
  Add support for a simple_server_retry_responses list (#6605)
  AuTest for server_push_preload plugin
  fix sed matching
  Improving the messaging around the use of TSSslSessionGetBuffer
  Fix PUSH_PROMISE frame payload length
  Cleanup: fix a inline function style
  Doc updates to TSContSchedule* API
  Skip compressible content type check with null strings.
  Include start line of HTTP messages in xdebug probe output.
  Convert mgmt/utils unit tests to use Catch.
  [Doc][DevGuide][TSStatSync] Fixing documentation for SUM and COUNT types
  Support body factory template suppression for internal requests
  Fix tunnel crash
  When using TSContSchedule() and TSContScheduleAPI() set the calling thread as the thread affinity when not already set
  Ensure TSContSchedule API family are called from an EThread.
  Make request/response body as an option for AuTest microserver
  Moved printing the incoming headers for debug before remapping
  SSL: Always renew TLS Session Tickets iff TLSv1.3 is being used
  Convert proxy/logging unit tests to use Catch.
  Convert single regression test in iocore/eventsystem/SocketManager.cc to Catch.
  Enable logging autests on macOS, clarify why Linux is required
  In Au tests, poll for excretion of log file rather than waiting fixed delay. (#6506)
  Fixed build issue with Ubuntu 16 debug
  Removed some unused declarations in HttpTransact
  Convert Mime and URL unit tests in proxy/hdrs to Catch.
  Convert unit tests for sslheaders plugin to Catch.
  free(map) -> delete map
  Removes copypasta curl text from tests and removes checks for curl
  Fixes some tls autests on macOS
  Add FetchSM support to dechunking in non-streaming mode. Add new TS API TSFetchFlagSet()
  Doc: Add mention for sleep time variance
  Allow lo interface in list of system stats (#6531)
  Remove unused unit test source files in src/tscore.
  better handling of TSVIO calls and TSVConnAbort (#6239)
  Convert tscore regression tests to Catch unit tests.
  Disables "virtual host not used with AWS auth v4" error in s3_auth
  Created remap_stats au test.  This tests 2 remaps, one for a 2xx response and one for a 4xx response and then checks the stats output (#6505)
  Adds support for configure option --enable-yaml-headers (#6519)
  Require 1.1.1 as minimum openssl lib version for tls_check_dual_cert_selection Au test.
  Removes noisy log statement from xdebug
  Make traffic_ctl limp along for now with large records.snap (#6517)
  Fixes a bug where getHostStatus() will create a host status record when none is found. A Host status record should only be created when a host is marked up or down when traffic_ctl is used to mark a host up or down.
  Moves hosting.config finished loading message outside of parsing loop
  Refactor and generalize the User Arg concept, add global (#6468)
  Check the exit value of the regression test
  Fix SDK_API_TSSslServerContextCreate
  Exponential backoff - Make values used configurable - Add config support for ceiling time and retry cap. - Add small variance when sleeping.
  Cleanup: check activity of Http2Stream by ActivityCop
  Rework server side SSL_CTX creation to better handle dual_cert mismatches (#6483)
  XDebug: Always append to the X- header, rather than prepend
  Doc: connection.match default value update
  Charge Proxy Header Regression tests into Catch unit tests.
  traffic_dump: Make client-request gathered in a global hook (#6500)
  Remove method that does nothing. (#6501)
  Doc: traffic_manager - Add documentation for exp backoff and configuration
  Make all_headers test more resilient to timimng
  remove pthread_cancel, use atomic flags to ensure cross thread safety, and some other cleanups
  Add null check to fix error tunnel crash
  LGTM: Fix comparing int8_t with wider type
  Fix a compile warning
  Bikeshedding some code structures for reloadable plugins config (#6488)
  Issue 3546: Add "overridable" to the configuration variable description.
  Remove some outdated files.
  Be explicit about RTLD_LOCAL, defaults varies on platforms (#6485)
  [CPPAPI] Provide access to TSRemapRequestInfo in RemapPlugins.

 Conflicts:
	iocore/net/quic/QUICTLS_boringssl.cc

+++ src/tscpp/util/TextView.cc
@@ -134,7 +134,7 @@ ts::svtoi(TextView src, TextView *out, int base)
       zret = svto_radix<16>(src);
       break;
     default:
-      while (src.size() && (0 <= (v = svtoi_convert[static_cast<unsigned char>(*src)])) && v < base) {
+      while (src.size() && (0 <= (v = svtoi_convert[static_cast<unsigned char>(*src)])) && static_cast<int>(v) < base) {
         auto n = zret * base + v;
         if (n < zret) {
           zret = std::numeric_limits<uintmax_t>::max();
